refactor: move filter, search, and sorting components to common folder
